refactor tests to separate files
