+++ tests/format.js
@@ -1,0 +1,16 @@
+/* jshint node:true */
+/* global describe,it */
+'use strict';
+var should = require('should');
+var fs = require('fs');
+
+var libPath = process.env.JSFMT_COV ? 'lib-cov' : 'lib';
+var jsfmt = require('../' + libPath + '/index');
+
+describe('jsfmt.format', function() {
+  it('should test basic formatting', function() {
+    var js = 'var func = function(test){console.log( test );};';
+    var result = jsfmt.format(js, {});
+    result.should.eql('var func = function(test) {\n  console.log(test);\n};');
+  });
+});
